HW2 debug 07/12/2022, should be right.

+++ hw2/hw2/PercolationStats.java
@@ -34,14 +34,11 @@ public class PercolationStats {
 
             }
 
-            xlist[i] = temp.numberOfOpenSites() / (N * N);
+            xlist[i] = temp.numberOfOpenSites() / (double)(N * N);
 
 
 
         }
-
-        //temp = pf.make(N);
-
 
         sizeN = N;
         timeT = T;
